requirements.txt
selenium==3.141.0
requests==2.25.1
webdriver-manager==3.8.6
beautifulsoup4==4.12.2
requests
python-multipart
scrapy==2.7.0
lxml~=4.9.2
Scrapy-Redis-BloomFilter==0.8.1
scrapy-redis==0.7.1
redis==3.0.1
pymongo==4.0.2
minio==7.1.14
Pillow==9.4.0
ujson
kafka-python==1.4.3
date_guesser==2.1.4
dateparser==1.1.5
feedparser~=6.0.2
boto3==1.26.99
python-dotenv==1.0.0
python-decouple
pandas
openpyxl